refactor(community): replace AuthHeader and AuthFooter with AuthLayout for improved structure
- Removed the AuthHeader and AuthFooter components, consolidating their functionality into a new AuthLayout component.
- Updated SignInPage and SignUpPage to utilize AuthLayout, enhancing readability and maintainability by centralizing layout logic.
- Introduced new styles for the AuthLayout to maintain visual consistency.
