+++ apps/community/src/app/(auth)/signin/page.tsx
@@ -1,17 +1,16 @@
-import { AuthFooter } from '~/features/auth/components/auth-footer'
-import { AuthHeader } from '~/features/auth/components/auth-header'
 import { SignInForm } from '~/features/auth/components/sign-in-form'
+import { AuthLayout } from '~/widgets/auth/components/auth-layout'
 
 export default function SignInPage() {
   return (
-    <>
-      <AuthHeader title="로그인" description="경기대학교 AI컴퓨터공학부" />
+    <AuthLayout
+      title="로그인"
+      description="경기대학교 AI컴퓨터공학부"
+      info="계정이 없으신가요?"
+      link="회원가입"
+      href="/signup"
+    >
       <SignInForm />
-      <AuthFooter
-        description="계정이 없으신가요?"
-        link="회원가입"
-        href="/signup"
-      />
-    </>
+    </AuthLayout>
   )
 }
