--- apps/community/src/features/auth/components/auth-footer.tsx
@@ -1,22 +1,0 @@
-import Link from 'next/link'
-
-import * as styles from '~/features/auth/components/auth-footer.css'
-
-interface Props {
-  description: string
-  link: string
-  href: string
-}
-
-function AuthFooter({ description, link, href }: Props) {
-  return (
-    <div className={styles.footerWrapper}>
-      <p>{description}</p>
-      <Link href={href} className={styles.link}>
-        {link}
-      </Link>
-    </div>
-  )
-}
-
-export { AuthFooter }
