+++ apps/community/src/widgets/auth/components/auth-layout.css.ts
@@ -2,6 +2,18 @@ import { style } from '@vanilla-extract/css'
 
 import { themeVars } from '@aics-client/design-system/styles'
 
+const headerWrapper = style({
+  textAlign: 'center',
+})
+
+const title = style([
+  themeVars.textSize['2xl'],
+  {
+    fontWeight: themeVars.fontWeight.bold,
+    marginBottom: themeVars.spacing.sm,
+  },
+])
+
 const footerWrapper = style({
   display: themeVars.display.flex,
   justifyContent: 'center',
@@ -20,4 +32,4 @@ const link = style({
   },
 })
 
-export { footerWrapper, link }
+export { headerWrapper, title, footerWrapper, link }
